+++ src/main/java/user/dao/DaoFactory.java
@@ -7,10 +7,10 @@ import org.springframework.context.annotation.Configuration;
 public class DaoFactory {
 
     @Bean
-    public UserDao userDao() {
-        UserDao userDao = new UserDao();
-        userDao.setConnectionMaker(connectionMaker());
-        return userDao;
+    public UserDaoJdbc userDao() {
+        UserDaoJdbc userDaoJdbc = new UserDaoJdbc();
+        userDaoJdbc.setConnectionMaker(connectionMaker());
+        return userDaoJdbc;
     }
 
     @Bean
