+++ src/main/java/user/dao/UserDao.java
@@ -1,59 +1,13 @@
 package user.dao;
 
-import org.springframework.dao.DuplicateKeyException;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
-import user.DuplicateUserIdException;
 import user.domain.User;
 
-import javax.sql.DataSource;
-import java.sql.ResultSet;
-import java.sql.SQLException;
 import java.util.List;
 
-public abstract class UserDao {
-
-    public void setDataSource(DataSource dataSource) {
-        this.jdbcTemplate = new JdbcTemplate(dataSource);
-    }
-
-    private JdbcTemplate jdbcTemplate;
-
-    private RowMapper<User> userMapper = new RowMapper<User>() {
-        @Override
-        public User mapRow(ResultSet rs, int rowNum) throws SQLException {
-            User user = new User();
-            user.setId(rs.getString("id"));
-            user.setName(rs.getString("name"));
-            user.setPassword(rs.getString("password"));
-            return user;
-        }
-    };
-
-    public void add(final User user) throws DuplicateUserIdException {
-        try {
-            this.jdbcTemplate.update("insert into users(id, name, password) values(?, ?, ?)",
-                    user.getId(), user.getName(), user.getPassword());
-        } catch (DuplicateKeyException e) {
-            throw new DuplicateUserIdException(e);
-        }
-    }
-
-    public void deleteAll() {
-        this.jdbcTemplate.update("delete from users");
-    }
-
-    public int getCount() {
-        return this.jdbcTemplate.queryForObject("select count(*) from users", Integer.class);
-    }
-
-    public User get(String id) {
-        return this.jdbcTemplate.queryForObject("select * from users where id = ?",
-                new Object[]{id}, this.userMapper);
-    }
-
-    public List<User> getAll() {
-        return this.jdbcTemplate.query("select * from users order by id",
-                this.userMapper);
-    }
+public interface UserDao {
+    void add(User user);
+    User get(String id);
+    List<User> getAll();
+    void deleteAll();
+    int getCount();
 }
