+++ src/test/java/UserDaoConnectionCountingTest.java
@@ -1,14 +1,14 @@
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 import user.dao.CountingConnectionMaker;
 import user.dao.CountingDaoFactory;
-import user.dao.UserDao;
+import user.dao.UserDaoJdbc;
 
 import java.sql.SQLException;
 
 public class UserDaoConnectionCountingTest {
     public static void main(String[] args) throws ClassNotFoundException, SQLException {
         AnnotationConfigApplicationContext context = new AnnotationConfigApplicationContext(CountingDaoFactory.class);
-        UserDao dao = context.getBean("userDao", UserDao.class);
+        UserDaoJdbc dao = context.getBean("userDao", UserDaoJdbc.class);
 
         //DAO 사용 코드
         CountingConnectionMaker ccm = context.getBean("connectionMaker", CountingConnectionMaker.class);
